requirements.txt
astroid==2.15.6
black==23.9.1
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
dill==0.3.7
google-api-core==2.11.1
google-auth==2.23.0
google-cloud-core==2.3.3
google-cloud-storage==2.11.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
idna==3.4
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
protobuf==4.24.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pylint==2.17.5
requests==2.31.0
rsa==4.9
tomlkit==0.12.1
urllib3==1.26.16
wrapt==1.15.0
